Disable ask_pass by default
Disable askpass by default (but enable it in `go.sh`). This makes it
possible to run user tasks only without having to do privileged stuff.

To make this not fail ansible needs the tags of all subtags in the play
with `become: true`. This makes no sense, but that's how it is.

Also add system and user tags for running all system and user tasks
respectively.